+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/Attributes.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 public class Attributes {
 
+    public List<String> uid;
+
     private List<String> localEntitlements;
 
     public List<String> getLocalEntitlements() {