+++ authorization/src/main/java/org/grnet/status/authorizations/dtos/GroupUserResponse.java
@@ -12,5 +12,7 @@ public class GroupUserResponse {
 
     public String email;
 
+    public String uid;
+
     public List<UserGroupInfoDto> tenants;
 }